+++ src/lt/martynassateika/idea/codeigniter/language/CiLanguageUtil.java
@@ -17,20 +17,20 @@
 package lt.martynassateika.idea.codeigniter.language;
 
 import com.intellij.openapi.project.Project;
+import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiManager;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.util.indexing.FileBasedIndex;
 import com.jetbrains.php.lang.PhpFileType;
 import com.jetbrains.php.lang.psi.PhpFile;
-import com.jetbrains.php.lang.psi.elements.ArrayAccessExpression;
-import com.jetbrains.php.lang.psi.elements.ArrayIndex;
 import com.jetbrains.php.lang.psi.elements.AssignmentExpression;
-import com.jetbrains.php.lang.psi.elements.PhpPsiElement;
+import com.jetbrains.php.lang.psi.elements.StringLiteralExpression;
 import com.jetbrains.php.lang.psi.visitors.PhpRecursiveElementVisitor;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import lt.martynassateika.idea.codeigniter.psi.MyPsiUtil;
 
 /**
  * Utility methods related to the CI Language class.
@@ -54,30 +54,44 @@ class CiLanguageUtil {
   }
 
   /**
+   * Returns {@code true} if the supplied element represents a language line key.
+   *
+   * @param element an element
+   * @return {@code true} if the supplied element represents a language line key
+   */
+  static boolean isLanguageLineKeyElement(PsiElement element) {
+    return MyPsiUtil.isArgumentOfFunction(element, "lang", 0);
+  }
+
+  /**
+   * @param project current project
+   * @param literalExpression containing the language key
+   * @return assignment expressions matching the language key across all language files
+   */
+  static List<AssignmentExpression> findTranslationsFor(
+      Project project,
+      StringLiteralExpression literalExpression) {
+    return findTranslationsFor(project, literalExpression.getContents());
+  }
+
+  /**
    * @param project current project
    * @param text language key
    * @return assignment expressions matching the language key across all language files
    */
-  @SuppressWarnings("deprecation")
-  static List<AssignmentExpression> findTranslationsFor(Project project, String text) {
+  private static List<AssignmentExpression> findTranslationsFor(Project project, String text) {
     List<AssignmentExpression> expressions = new ArrayList<>();
     FileBasedIndex.getInstance().getFilesWithKey(LanguageFileIndex.KEY,
         Collections.singleton(text), file -> {
           PsiManager psiManager = PsiManager.getInstance(project);
           PsiFile psiFile = psiManager.findFile(file);
           if (psiFile instanceof PhpFile) {
+            //noinspection deprecation
             psiFile.accept(new PhpRecursiveElementVisitor() {
               @Override
               public void visitPhpAssignmentExpression(AssignmentExpression assignmentExpression) {
-                PhpPsiElement variable = assignmentExpression.getVariable();
-                if (variable instanceof ArrayAccessExpression) {
-                  ArrayAccessExpression arrayAccessExpression = (ArrayAccessExpression) variable;
-                  ArrayIndex index = arrayAccessExpression.getIndex();
-                  if (index != null) {
-                    if (index.getText().equals(text)) {
-                      expressions.add(assignmentExpression);
-                    }
-                  }
+                if (MyPsiUtil.isArrayAccessWithStringIndex(assignmentExpression, text)) {
+                  expressions.add(assignmentExpression);
                 }
               }
             });